renamed CLIAdapter to ShellWrapper, added interface to inherit from
