+++ ShellAdapter/IShell.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace ShellAdapter
+{
+	public interface IShell
+	{
+		public string Execute(string command, string[] args);
+	}
+}
